+++ spec/dummy/client/app/components/HelloWorldRedux.jsx
@@ -38,7 +38,7 @@ export default class HelloWorldRedux extends React.Component {
           <input
             type="text"
             ref={this.setNameDomRef}
-            defaultValue={name}
+            value={name}
             onChange={this.handleChange}
           />
         </p>