+++ spec/dummy/config/routes.rb
@@ -6,6 +6,8 @@ Rails.application.routes.draw do
   root "pages#index"
 
   get "client_side_hello_world" => "pages#client_side_hello_world"
+  get "client_side_hello_world_shared_store" => "pages#client_side_hello_world_shared_store"
+  get "server_side_hello_world_shared_store" => "pages#server_side_hello_world_shared_store"
   get "server_side_hello_world" => "pages#server_side_hello_world"
   get "client_side_log_throw" => "pages#client_side_log_throw"
   get "server_side_log_throw" => "pages#server_side_log_throw"